Fixed comment "//todo wrong package por util class" by moving the class to Utils package
